add tests for ContiguousStorage
